Implement permuteAndSort to statistic max_depth
The function permuteAndSort will find all permutations of a given array
and use each permutation to build a list then pass the list to
quick_sort function. This can be convinient for us to statistic the
possible maximum depth for begin[] array.

Mind that this function assume every element in the array is unique.