Use head->prev to store tail of each sublist
We didn't need to use prev link to traverse the link in quicksort, thus
we can store the tail info of each sublist in it. By doing so, the time
complexity of finding the list's tail node if decresed from O(n) into
O(1), where n is the node cnt of given list.