Implement of powerSort
The Power sort is another variant of timsort, whose main idea is to
imagine a balanced binary tree and try to make the lengths of the
unmerged sublists as close as possible to the lengths of each subtree
in the balanced binary tree. This ensures as much balance as possible
when merging sublists.